requirements.txt
# Core dependencies (download + train + signal + backtest + dashboard)
pyqlib>=0.9.0
lightgbm>=4.0.0
xgboost>=2.0.0
pandas>=2.0.0
numpy>=2.0,<2.5
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
streamlit>=1.28.0
pyyaml>=6.0.0
loguru>=0.7.0
schedule>=1.2.0
markdown2>=2.4.0
click>=8.1.0
joblib>=1.3.0
requests>=2.31.0
tenacity>=8.2.0
python-dateutil>=2.8.0
akshare>=1.14.0
baostock>=0.8.8
efinance>=0.5.0
tushare>=1.4.0
litellm>=1.0.0
json_repair>=0.1.0

# Web API
fastapi>=0.110.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0

# Search & news (used by search_service.py, graceful degradation if missing)
newspaper3k>=0.2.8
lxml_html_clean>=0.4.0
tavily-python>=0.3.0
google-search-results>=2.4.2

# Optional: uncomment if needed
# torch>=2.0.0          # MLP/Transformer models (~900MB)
# vnpy>=3.9.0           # Live trading engine
# vnpy-xtp>=2.2.0       # XTP broker gateway
# redis>=5.0.0          # Real-time signal pub/sub
# pymysql>=1.1.0        # MySQL driver
# ta-lib>=0.4.28        # TA-Lib (requires C library)
# seaborn>=0.12.0       # Extra plotting
# plotly>=5.18.0        # Interactive charts
